+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Possible log types:
 ### v0.3.1 (UNRELEASED)
 
 - [added] Top level `get_version` function (#31)
+- [added] Added `Status::ext_versions` field (#31)
 - ...
 
 ### v0.3.0 (2016-06-25)